+++ src/ByteString/Aeson/Orphans.hs
@@ -37,28 +37,3 @@ instance FromJSON LBS.ByteString where
 
 instance ToJSONKey ByteString
 instance FromJSONKey ByteString
-
-#if !MIN_VERSION_aeson(1,0,2)
-deriving instance FromJSON a => FromJSON (First a)
-deriving instance ToJSON a => ToJSON (First a)
-#endif
-
-deriving instance FromJSON Any
-deriving instance ToJSON Any
-
-instance (ToJSON (f a)) => ToJSON (Alt f a)
-instance (FromJSON (f a)) => FromJSON (Alt f a)
-
-#if !MIN_VERSION_aeson(1,0,2)
-instance (Ord k, FromJSON k, FromJSON v) => FromJSON (Map k v) where
-  parseJSON = parseJSONMap
-
-instance (ToJSON k, ToJSON v) => ToJSON (Map k v) where
-  toJSON = toJSONMap
-#endif
-
-parseJSONMap :: (Ord k, FromJSON k, FromJSON v) => Value -> Parser (Map k v)
-parseJSONMap v = Map.fromList <$> parseJSON v
-
-toJSONMap :: (ToJSON k, ToJSON v) => Map k v -> Value
-toJSONMap = toJSON . Map.toList